Move ConfigToolBase to PhysicsTools/PatAlgos
