+++ PhysicsTools/PatAlgos/python/tools/cmsswVersionTools.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 from builtins import range
 import FWCore.ParameterSet.Config as cms
 
-from FWCore.Utilities.ConfigToolBase import *
+from PhysicsTools.PatAlgos.tools.ConfigToolBase import *
 from PhysicsTools.PatAlgos.tools.helpers import *
 from PhysicsTools.PatAlgos.tools.jetTools import *
 from Configuration.AlCa.autoCond import autoCond
