+++ PhysicsTools/PatAlgos/python/tools/coreTools.py
@@ -1,6 +1,6 @@
 from __future__ import print_function
 from builtins import range
-from FWCore.Utilities.ConfigToolBase import *
+from PhysicsTools.PatAlgos.tools.ConfigToolBase import *
 
 from PhysicsTools.PatAlgos.tools.helpers import *
