+++ PhysicsTools/PatUtils/python/tools/eGammaCorrection.py
@@ -1,7 +1,7 @@
 from __future__ import print_function
 import FWCore.ParameterSet.Config as cms
 
-from FWCore.Utilities.ConfigToolBase import *
+from PhysicsTools.PatAlgos.tools.ConfigToolBase import *
 #import PhysicsTools.PatAlgos.tools.helpers as configtools
 
 from PhysicsTools.PatAlgos.tools.helpers import getPatAlgosToolsTask, addToProcessAndTask
