+++ PhysicsTools/PatUtils/python/tools/muonRecoMitigation.py
@@ -1,6 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 
-from FWCore.Utilities.ConfigToolBase import *
+from PhysicsTools.PatAlgos.tools.ConfigToolBase import *
 
 from PhysicsTools.PatAlgos.tools.helpers import getPatAlgosToolsTask, addToProcessAndTask
