+++ PhysicsTools/PatUtils/python/tools/runJetUncertainties.py
@@ -1,6 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 
-from FWCore.Utilities.ConfigToolBase import *
+from PhysicsTools.PatAlgos.tools.ConfigToolBase import *
 
 import PhysicsTools.PatAlgos.tools.helpers as configtools
 from PhysicsTools.PatAlgos.tools.trigTools import _addEventContent
